% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot.R
\name{function_plot}
\alias{function_plot}
\title{function_plot is for making plot based on the result of myseq_n.}
\usage{
function_plot(df)
}
\arguments{
\item{df}{}
}
\value{
plot
}
\description{
function_plot is for making plot based on the result of myseq_n.
}
\examples{
my_data function_plot(my_data)
}
