% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pyanzhang7559resub-package.R
\docType{package}
\name{hw04pyanzhang7559resub-package}
\alias{hw04pyanzhang7559resub}
\alias{hw04pyanzhang7559resub-package}
\title{hw04pyanzhang7559resub: This is for homework 4 builing a package}
\description{
This is calculate a n order number based on the first, second and third element in a vector. Then you can make a plot based on the vector.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yanzhang7559/hw04pyanzhang7559resub}
  \item Report bugs at \url{https://github.com/yanzhang7559/hw04pyanzhang7559resub/issues}
}

}
\author{
\strong{Maintainer}: Yan Zhang \email{yz7559a@student.american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
