% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_n.R
\name{myseq_n}
\alias{myseq_n}
\title{myseq_n is for calculating the nth number based on the first 3 numbers.}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{the vector in the input}

\item{n}{the order of n need to be calculate}
}
\value{
my_seq_n
}
\description{
myseq_n is for calculating the nth number based on the first 3 numbers.
}
\examples{
myseq_n(x = c(2, 4, 3), n = 4)
}
